Create cluster info once, per server, regardless of system creation
