+++ src/server/system_services/system_server.js
@@ -10,7 +10,6 @@ const url = require('url');
 const net = require('net');
 const moment = require('moment');
 const ip_module = require('ip');
-const uuid = require('node-uuid');
 const P = require('../../util/promise');
 const pkg = require('../../../package.json');
 const dbg = require('../../util/debug_module')(__filename);
@@ -89,18 +88,6 @@ function new_system_changes(name, owner_account) {
                 system: system._id,
                 role: 'admin'
             };
-            var cluster = {
-                owner_secret: system_store.get_server_secret(),
-                cluster_id: uuid().substring(0, 8),
-                shards: [{
-                    shardname: 'shard1',
-                    servers: [{
-                        address: os_utils.get_local_ipv4_ips()[0] //TODO:: on multiple nics support, fix this
-                    }],
-                }],
-                config_servers: [],
-            };
-
             ActivityLog.create({
                 event: 'conf.create_system',
                 level: 'info',
@@ -117,7 +104,6 @@ function new_system_changes(name, owner_account) {
                     tiers: [tier],
                     pools: [pool],
                     roles: [role],
-                    clusters: [cluster],
                 }
             };
